Class to get specific query from Folder
