+++ app/models/folders/subfolder.rb
@@ -1,7 +1,7 @@
 module Folders
   class Subfolder < Folder
     validates :name, uniqueness: true, if: proc {
-      SubfoldersQuery.new(name: name, parent_id: parent_id).exists?
+      FoldersQuery.new(name: name, parent_id: parent_id).exists?
     }
   end
 end
